회원 탈퇴시 작성한 글 먼저 삭제

+++ src/main/java/com/example/prj1be1109/Controller/BoardController.java
@@ -4,14 +4,12 @@ package com.example.prj1be1109.Controller;
 import com.example.prj1be1109.Service.BoardService;
 import com.example.prj1be1109.domain.Board;
 import com.example.prj1be1109.domain.Member;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Objects;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,9 +19,9 @@ public class BoardController {
     private final BoardService service;
 
     @PostMapping("add")
-    public ResponseEntity add(@RequestBody Board board, @SessionAttribute(value = "login", required = false) Member login) {
+    public ResponseEntity add(@RequestBody Board board,
+                              @SessionAttribute(value = "login", required = false) Member login) {
 
-        System.out.println("login" + login);
 //        로그인 값이 null이라면
         if(login == null){
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
@@ -50,7 +48,14 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute(value = "login", required = false)Member login) {
+        if (login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401
+        }
+        if (!service.hasAccess(id,login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
         if (service.remove(id)) {
             return ResponseEntity.ok().build();
         } else {
@@ -59,7 +64,14 @@ public class BoardController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity<Object> edit(@RequestBody Board board) {
+    public ResponseEntity<Object> edit(@RequestBody Board board,
+                                       @SessionAttribute(value ="login",required = false) Member login) {
+        if (login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); //401
+        }
+        if (!service.hasAccess(board.getId(),login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); //403
+        }
         if (service.validate(board)) {
             if (service.update(board)) {
                 return ResponseEntity.ok().build();
